% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nivowaffle.R
\name{nivowaffle-shiny}
\alias{nivowaffle-shiny}
\alias{nivowaffleOutput}
\alias{renderNivowaffle}
\alias{nivowaffle_html}
\title{Shiny bindings for nivowaffle}
\usage{
nivowaffleOutput(outputId, width = "100\%", height = "400px")

renderNivowaffle(expr, env = parent.frame(), quoted = FALSE)

nivowaffle_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a nivowaffle}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using nivowaffle within Shiny
applications and interactive Rmd documents.
}
