% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nivowaffle.R
\name{nivowaffle}
\alias{nivowaffle}
\title{绘制华夫图}
\usage{
nivowaffle(data = NULL, total = 100, rows = 18, columns = 14,
  colorBy = "id", ..., width = NULL, height = NULL,
  elementId = NULL)
}
\description{
绘制华夫图
}
